# backend/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Travellle API
  description: REST API for Travellle travel planning app
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
tags:
  - name: plans
    description: Travel plan operations
  - name: pins
    description: Pin operations on plans
  - name: notes
    description: Note operations on pins
  - name: spots
    description: Spot search operations
paths:
  /plans:
    post:
      tags:
        - plans
      summary: Create a new travel plan
      operationId: createPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlanRequest'
      responses:
        '201':
          description: Plan created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /plans/{share_token}:
    get:
      tags:
        - plans
      summary: Fetch a plan by sharing URL token
      operationId: getPlanByShareToken
      parameters:
        - name: share_token
          in: path
          required: true
          schema:
            type: string
          description: Unique token used in the sharing URL
          example: "abc123xyz"
      responses:
        '200':
          description: Plan found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '404':
          description: Plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /plans/{plan_id}/pins:
    get:
      tags:
        - pins
      summary: List all pins for a plan
      operationId: listPins
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Target plan ID
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Pins retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pin'
        '400':
          description: Invalid plan_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - pins
      summary: Add a pin to a plan
      operationId: createPin
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Target plan ID
          example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePinRequest'
      responses:
        '201':
          description: Pin added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pin'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /plans/{plan_id}/pins/{pin_id}:
    patch:
      tags:
        - pins
      summary: Update pin category and/or colour
      operationId: updatePin
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Target plan ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        - name: pin_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Target pin ID
          example: "660e8400-e29b-41d4-a716-446655440001"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePinRequest'
      responses:
        '200':
          description: Pin updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pin'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Plan or Pin not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - pins
      summary: Delete a pin from a plan
      operationId: deletePin
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Target plan ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        - name: pin_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Target pin ID
          example: "660e8400-e29b-41d4-a716-446655440001"
      responses:
        '204':
          description: Pin deleted successfully
        '404':
          description: Plan or Pin not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /plans/{plan_id}/pins/{pin_id}/notes:
    post:
      tags:
        - notes
      summary: Add a note to a pin
      operationId: createNote
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Target plan ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        - name: pin_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Target pin ID
          example: "660e8400-e29b-41d4-a716-446655440001"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNoteRequest'
      responses:
        '201':
          description: Note added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Pin not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /plans/{plan_id}/pins/{pin_id}/notes/{note_id}:
    patch:
      tags:
        - notes
      summary: Update a note
      operationId: updateNote
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Target plan ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        - name: pin_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Target pin ID
          example: "660e8400-e29b-41d4-a716-446655440001"
        - name: note_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Target note ID
          example: "770e8400-e29b-41d4-a716-446655440002"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNoteRequest'
      responses:
        '200':
          description: Note updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Pin or Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - notes
      summary: Delete a note from a pin
      operationId: deleteNote
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Target plan ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        - name: pin_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Target pin ID
          example: "660e8400-e29b-41d4-a716-446655440001"
        - name: note_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Target note ID
          example: "770e8400-e29b-41d4-a716-446655440002"
      responses:
        '204':
          description: Note deleted successfully
        '404':
          description: Pin or Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /spots/search:
    get:
      tags:
        - spots
      summary: Search spots using Google Places API
      operationId: searchSpots
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
          description: Search keyword for spot lookup
          example: "Tokyo Tower"
      responses:
        '200':
          description: Search results returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Spot'
        '400':
          description: Missing or invalid query parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error or Google Places API error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: "an error occurred"
    Plan:
      type: object
      required:
        - id
        - share_token
        - title
        - pins
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        share_token:
          type: string
          example: "abc123xyz"
        title:
          type: string
          example: "Tokyo Trip 2025"
        pins:
          type: array
          items:
            $ref: '#/components/schemas/PinWithNotes'
        created_at:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
    CreatePlanRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          example: "Tokyo Trip 2025"
    Pin:
      type: object
      required:
        - id
        - plan_id
        - name
        - latitude
        - longitude
        - category
        - colour
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          example: "660e8400-e29b-41d4-a716-446655440001"
        plan_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          example: "Tokyo Tower"
        latitude:
          type: number
          format: double
          example: 35.6586
        longitude:
          type: number
          format: double
          example: 139.7454
        category:
          type: string
          enum:
            - restaurant
            - hotel
            - sightseeing
            - transport
            - other
          example: "sightseeing"
        colour:
          type: string
          example: "#FF5733"
        created_at:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
    PinWithNotes:
      allOf:
        - $ref: '#/components/schemas/Pin'
        - type: object
          required:
            - notes
          properties:
            notes:
              type: array
              items:
                $ref: '#/components/schemas/Note'
    CreatePinRequest:
      type: object
      required:
        - name
        - latitude
        - longitude
        - category
        - colour
      properties:
        name:
          type: string
          example: "Tokyo Tower"
        latitude:
          type: number
          format: double
          example: 35.6586
        longitude:
          type: number
          format: double
          example: 139.7454
        category:
          type: string
          enum:
            - restaurant
            - hotel
            - sightseeing
            - transport
            - other
          example: "sightseeing"
        colour:
          type: string
          example: "#FF5733"
    UpdatePinRequest:
      type: object
      properties:
        category:
          type: string
          enum:
            - restaurant
            - hotel
            - sightseeing
            - transport
            - other
          example: "restaurant"
        colour:
          type: string
          example: "#3498DB"
    Note:
      type: object
      required:
        - id
        - pin_id
        - content
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          example: "770e8400-e29b-41d4-a716-446655440002"
        pin_id:
          type: string
          format: uuid
          example: "660e8400-e29b-41d4-a716-446655440001"
        content:
          type: string
          example: "Great view from the top!"
        created_at:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
    CreateNoteRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          example: "Great view from the top!"
    UpdateNoteRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          example: "Updated note content"
    Spot:
      type: object
      required:
        - place_id
        - name
        - address
        - latitude
        - longitude
      properties:
        place_id:
          type: string
          description: Google Places API place ID
          example: "ChIJ51cu8IcbXWARiRtXIothAS4"
        name:
          type: string
          example: "Tokyo Tower"
        address:
          type: string
          example: "4 Chome-2-8 Shibakoen, Minato City, Tokyo 105-0011"
        latitude:
          type: number
          format: double
          example: 35.6586
        longitude:
          type: number
          format: double
          example: 139.7454